% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{addToMetaData}
\alias{addToMetaData}
\title{Add field to meta data}
\usage{
addToMetaData(corpus, vector, tag)
}
\arguments{
\item{corpus}{corpus object}

\item{vector}{vector of values to add to corpus. must be of equal length to corpus}

\item{tag}{meta data tag where values are to be added}
}
\description{
Add field to meta data
}

