% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{characterCount}
\alias{characterCount}
\title{Character Count}
\usage{
characterCount(corp)
}
\arguments{
\item{corp}{Corpus object}
}
\description{
Count the number of characters in each document in a given corpus, and return as a vector
}

