% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{contextAnalyser}
\alias{contextAnalyser}
\title{Search Context Analyser}
\usage{
contextAnalyser(corpus, terms, search.vector = NA, width = 0)
}
\description{
Search Context Analyser
}

