% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xts_functions.R
\name{corpusAggregate}
\alias{corpusAggregate}
\title{Apply function over corpus}
\usage{
corpusAggregate(corpus, meta_time, time_aggregate, FUN, ...)
}
\arguments{
\item{corpus}{corpus object over which to aggregate}

\item{meta_time}{meta data field containing time information}

\item{time_aggregate}{period over which to aggregate. Can be one of the following: \{day,week,month,quarter,year\}}

\item{FUN}{function to apply to corpus time bins}

\item{...}{optional arguments to FUN}
}
\description{
Apply function over corpus within specified time periods, and return an aggregated xts object.
For example, the following function calculates the number of unique terms in a given corpus:
\code{FUN<-function(x){nTerms(DocumentTermMatrix(x))}}
\code{corpusAggregate} splits the supplied corpus in to time periods, then applies the function to each
individual period in isolation.
}

