% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{corpusProportion}
\alias{corpusProportion}
\title{Corpus Proportion}
\usage{
corpusProportion(mat)
}
\arguments{
\item{mat}{matrix of frequency scores}
}
\description{
Given a matrix of frequency scores return the proportion of documents that matched each term
in the matrix.
}

