% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{extractMetaData}
\alias{extractMetaData}
\title{Extract meta data}
\usage{
extractMetaData(corpus)
}
\arguments{
\item{corpus}{corpus object with associated meta data}
}
\description{
Extract meta data in to a data frame from a corpus object
}

