% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{loadMetaData}
\alias{loadMetaData}
\title{Load meta data on to a corpus}
\usage{
loadMetaData(corpus, metadata)
}
\arguments{
\item{corpus}{tm corpus object}

\item{list}{of lists, where each higher level list item is a named meta data field,
and each lower level list is of equal length to the corpus, and contains values for the given field. `extractMetaData()` produces
compatible output.}
}
\description{
Load meta data on to a corpus
}

