% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{logicalMatch}
\alias{logicalMatch}
\title{Logical concatenation}
\usage{
logicalMatch(search, logic, FUN = NULL)
}
\arguments{
\item{search}{matrix of search results. Each column of results for a different term.}

\item{logic}{value
#'\itemize{
 \item{AND - if all terms matched, return count}
 \item{OR - return all term counts}
}}

\item{FUN}{either a character vector indicating to return a logical vector of matches,
or a function specifying the normalisation of matching values. For example, if normalisation = sum, for an "OR"
logical match all matches of each search term in each document will be summed and returned. If normalisation = min,
for an "AND" logical match the number of pairs of terms is returned.}
}
\description{
Concatenates rows of a search result matrix  according to a logical condition.
}

