% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO_functions.R, R/narrative_functions.R, R/xts_functions.R
\docType{package}
\name{Narrative}
\alias{Narrative}
\alias{Narrative-package}
\title{Read and write to and from corpus}
\description{
Read and write to and from corpus

Narrative corpus manipulation functions

Narrative xts functions
}

