% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{narrativeAnnotator}
\alias{narrativeAnnotator}
\title{Annotate corpus}
\usage{
narrativeAnnotator(corpus, annotator = "sentence", meta = T)
}
\description{
Apply annotations to documents.
}

