% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{readSeparateText}
\alias{readSeparateText}
\title{Read separate text documents within a given directory, output corpus.}
\usage{
readSeparateText(directory, load_metadata = F, meta_directory = directory,
  metadata_filename = NULL)
}
\arguments{
\item{directory}{path to text file directory}
}
\description{
Read separate text documents within a given directory, output corpus.
}

