% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{tdmGenerator}
\alias{tdmGenerator}
\title{Generate Term Document Matrices}
\usage{
tdmGenerator(length, corpus, control_params = c())
}
\arguments{
\item{length}{character vector of terms, or a numeric vector of lengths, for which to create corresponding term document matrices}

\item{corpus}{corpus object}
}
\description{
Generate appropriate order term document matrices for a given list of terms or numeric values.
}

