% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{timeNormalise}
\alias{timeNormalise}
\title{(DEPRECATED) Normalisation Generator}
\usage{
timeNormalise(corpus, meta_time, time_aggregate, date.sequence = NULL)
}
\arguments{
\item{corpus}{corpus object}

\item{meta_time}{document level meta data field that defines time}

\item{time_aggregate}{time resolution to aggregate over. Can be one of the following:
\itemize{
 \item{none - just returns un aggregated data, with NAs filtered}
 \item{daily}
 \item{weekly}
 \item{monthly}
 \item{quarterly}
 \item{yearly}
}}
}
\description{
Generates a data frame of common normalisation metrics
(term, character, word and document counts) for a given corpus
over aggregated time periods.
}

