% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{timeSeriesNarrative}
\alias{timeSeriesNarrative}
\title{Aggregated Time Series Objects by Meta Data}
\usage{
timeSeriesNarrative(corpus, meta_time, time_aggregate, meta_field,
  meta_normalisation_field = F, aggregate_function = sum)
}
\arguments{
\item{corpus}{corpus to plot over}

\item{meta_time}{document level meta data field that defines time}

\item{time_aggregate}{time resolution to aggregate over. Can be one of the following:
\itemize{
 \item{none - just returns un aggregated data, with NAs filtered}
 \item{daily}
 \item{weekly}
 \item{monthly}
 \item{quarterly}
 \item{yearly}
}}

\item{meta_field}{document level meta data field over which to aggregate}

\item{aggregate_function}{*optional* function to apply to aggregated fields. default=sum}
}
\description{
Returns an xts time series object aggregated over the specified time window.
NAs are implicitly removed during creation of the xts object
}

