% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{weightSort}
\alias{weightSort}
\title{Weight a Term Document Matrix}
\usage{
weightSort(x)
}
\arguments{
\item{corpus}{corpus object to be weighted}
}
\description{
Returns a nested list of terms for each document, sorted by weight.
}

