% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{wordCount}
\alias{wordCount}
\title{Word Count}
\usage{
wordCount(x)
}
\arguments{
\item{x}{DocumentTermMatrix or TermDocumentMatrix object, from the tm package.}
}
\description{
Count the number of words in each document in a given corpus, and return as a vector
}

