% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xts_functions.R
\name{xtsGenerate}
\alias{xtsGenerate}
\title{Generate an xts object}
\usage{
xtsGenerate(time, value)
}
\arguments{
\item{time}{data frame of time values}

\item{value}{data frame of values corresponding to times}
}
\description{
Returns an xts object given time and value data frames
}

