% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{corpusSentiment}
\alias{corpusSentiment}
\title{Financial Narratives Sentiment}
\usage{
corpusSentiment(tdm, dict.positive, dict.negative, normalisation.meta = NULL)
}
\arguments{
\item{tdm}{TermDocumentMatrix object}

\item{dict.positive}{character vector of positive words to be matched}

\item{dict.negative}{character vector of negative words to be matched}

\item{normalisation.meta}{document level corpus meta data field containing value to normalise over.}
}
\description{
#Calculate sentiment of documents in a corpus using the sentiment calculator from Financial Narratives.
}

