% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{narrativeAnnotator}
\alias{narrativeAnnotator}
\title{Annotate corpus}
\usage{
narrativeAnnotator(corpus, annotator = "sentence", metadata = T)
}
\arguments{
\item{corpus}{tm Corpus object}

\item{annotator}{specify "sentence" or "word" tokenisation, or provide your own annotator function}
}
\description{
Apply annotations to documents.
}

