% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{readSeparateText}
\alias{readSeparateText}
\title{Read separate text documents within a given directory, output corpus.}
\usage{
readSeparateText(directory, load_metadata = F, meta_directory = directory,
  metadata_filename = NULL)
}
\arguments{
\item{directory}{path to text file directory}

\item{load_metadata}{bool, whether to load metadata along with the corpus or not}

\item{meta_directory}{if load_metadata is True, specify the location directory. Defaults to `directory`.}

\item{metadata_filename}{if load_metadata is True, specify the filename.}
}
\description{
Read separate text documents within a given directory, output corpus.
}

