% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IO_functions.R
\name{saveCorpus}
\alias{saveCorpus}
\title{Save Corpus}
\usage{
saveCorpus(corpus, directory, filenames = NULL, save_metadata = F,
  meta_directory = directory)
}
\arguments{
\item{corpus}{a tm corpus object}

\item{directory}{ouput directory}

\item{filenames}{filenames with which to save each document. If `NULL`, defaults to document id metadata in corpus.}

\item{save_metadata}{switch to indicate that metadata should be saved}

\item{meta_directory}{ouput metadata directory. defaults to `dir`}
}
\description{
Save with meta data
}

