% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/narrative_functions.R
\name{weightSort}
\alias{weightSort}
\title{Sort a Term Document Matrix by term weight}
\usage{
weightSort(x)
}
\arguments{
\item{x}{document term or term document matrix object to be weighted}
}
\description{
Returns a nested list of terms for each document, sorted by weight.
}

