% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xts_functions.R
\name{xtsAggregate}
\alias{xtsAggregate}
\title{Aggregate xts objects}
\usage{
xtsAggregate(xts.scores, time_aggregate, normalisation,
  aggregate_function = sum)
}
\description{
Returns an xts object aggregated and normalised over the given time window by the specified aggregation function
}

